encapsulation of chunk functionality
